Move image deleting tests to separate spec
